+++ src/joyride/getting_started.cljs
@@ -24,49 +24,24 @@
     (js/console.info "Creating " ^String (.-fsPath dest-uri))
     (vscode/workspace.fs.copy source-uri dest-uri)))
 
-(defn user-activate-uri-section-and-subpath []
-  [(conf/user-abs-scripts-path)
-   "user"
-   "user_activate.cljs"])
-
-(defn user-hello-uri-section-and-subpath []
-  [(conf/user-abs-scripts-path)
-   "user"
-   "hello_joyride_user_script.cljs"])
-
-(defn user-my-lib-uri-section-and-subpath []
-  [(conf/user-abs-scripts-path)
-   "user"
-   "my_lib.cljs"])
-
-(defn workspace-activate-uri-section-and-subpath []
-  [(conf/workspace-abs-scripts-path)
-   "workspace"
-   "workspace_activate.cljs"])
-
-(defn workspace-hello-uri-section-and-subpath []
-  [(conf/workspace-abs-scripts-path)
-   "workspace" 
-   "hello_joyride_workspace_script.cljs"])
-
 (defn- dest-uri-uri-exists?+ [section-and-subpath]
   (p/let [dest-uri (apply destination-uri section-and-subpath)
           exists?+ (utils/path-or-uri-exists?+ dest-uri)]
     [dest-uri exists?+]))
 
-(defn maybe-create-user-content+ []
-  (p/let [section-and-subpath (user-activate-uri-section-and-subpath)
-          [activate-dest-uri activate-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
+(defn- maybe-create-content+ [section-and-subpath]
+  (p/let [[activate-dest-uri activate-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
     (when-not activate-exists?+
-      (apply (partial create-content-file+ activate-dest-uri) section-and-subpath)))
-  (p/let [section-and-subpath (user-hello-uri-section-and-subpath)
-          [hello-dest-uri hello-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
-    (when-not hello-exists?+ 
-      (apply (partial create-content-file+ hello-dest-uri) section-and-subpath)))
-  (p/let [section-and-subpath (user-my-lib-uri-section-and-subpath)
-          [my-lib-dest-uri my-lib-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
-    (when-not my-lib-exists?+ 
-      (apply (partial create-content-file+ my-lib-dest-uri) section-and-subpath))))
+      (apply (partial create-content-file+ activate-dest-uri) section-and-subpath))))
+
+(defn maybe-create-user-content+ []
+  (p/let [[_ user-activate-exists?] (dest-uri-uri-exists?+ [(conf/user-abs-scripts-path) "user" "user_activate.cljs"])]
+    (when-not user-activate-exists?
+      (maybe-create-content+ [(conf/user-abs-scripts-path) "user" "user_activate.cljs"])
+      (maybe-create-content+ [(conf/user-abs-scripts-path) "user" "hello_joyride_user_script.cljs"])
+      (maybe-create-content+ [(conf/user-abs-scripts-path) "user" "hello_joyride_user_script.js"])
+      (maybe-create-content+ [(conf/user-abs-src-path) "user" "problem_hover.cljs"])
+      (maybe-create-content+ [(conf/user-abs-src-path) "user" "my_lib.cljs"]))))
 
 (defn create-and-open-content-file+ [content-file-uri section-and-subpath]
   (fn []
@@ -75,14 +50,13 @@
                 (vscode/window.showTextDocument
                  #js {:preview false, :preserveFocus false})))))
 
-(defn maybe-create-workspace-activate-fn+ []
-  (p/let [section-and-subpath (workspace-activate-uri-section-and-subpath)
-          [activate-dest-uri activate-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
+(defn maybe-create-and-open-content+ [section-and-subpath]
+  (p/let [[activate-dest-uri activate-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
     (when-not activate-exists?+
       (create-and-open-content-file+ activate-dest-uri section-and-subpath))))
 
+(defn maybe-create-workspace-activate-fn+ []
+  (maybe-create-and-open-content+ [(conf/workspace-abs-scripts-path) "workspace" "workspace_activate.cljs"]))
+
 (defn maybe-create-workspace-hello-fn+ []
-  (p/let [section-and-subpath (workspace-hello-uri-section-and-subpath)
-          [hello-dest-uri hello-exists?+] (dest-uri-uri-exists?+ section-and-subpath)]
-    (when-not hello-exists?+
-      (create-and-open-content-file+ hello-dest-uri section-and-subpath))))
+  (maybe-create-and-open-content+ [(conf/workspace-abs-scripts-path) "workspace" "hello_joyride_workspace_script.cljs"]))